LayoutTests/platform/mac-ventura-wk2-lbse-text/svg/as-object/embedded-svg-size-changes-expected.txt


This test checks that intrinsic size changes of an embedded SVG cause immediate updates of the object that embeds it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check initial <object> size, before external resource loaded
PASS document.defaultView.getComputedStyle(object1).width is "300px"
PASS document.defaultView.getComputedStyle(object1).height is "150px"

Check initial <object> size, after external resource loaded
FAIL document.defaultView.getComputedStyle(object1).width should be 200px. Was 300px.
FAIL document.defaultView.getComputedStyle(object1).height should be 200px. Was 150px.

Check final <object> size, after resizing finished
FAIL document.defaultView.getComputedStyle(object1).width should be 400px. Was 300px.
FAIL document.defaultView.getComputedStyle(object1).height should be 200px. Was 150px.

Test passed if you see two green rectangles
PASS successfullyParsed is true

TEST COMPLETE
